+++ scripts/lichessTools.js
@@ -187,6 +187,8 @@
       Logo: '\ue07a',
       Switch: '\ue07b',
       Blindfold: '\ue07c',
+      NextLine: '\ue07d',
+      Stockfish: '\ue07e',
 
       // LiChess Tools icons
       ShowTranspositions: 'T',
@@ -227,6 +229,7 @@
       SmilingFaceWithHorns: '\uD83D\uDE08',
       SparklingHeart: '\uD83D\uDC96',
       SuperscriptCapitalL: '\u00a0\u1D38',
+      SuperscriptCapitalT: '\u00a0\u1D40',
       WhiteChessPawn: '\u2659',
       Opposition: '\u260D',
       Document: '\uD83D\uDDCE',