+++ tools/extraChart/tool.js
@@ -680,11 +680,10 @@
       if (mwStartUci * side + 1 + delta < mwEndUci * side) {
         return 1 + bonus;
       }
+      const mmw1 = this.maxMaterialWon(board, side) / 100;
       board = lt.getBoardFromFen(prev2Node.fen);
       const mmw3 = this.maxMaterialWon(board, side) / 100;
-      board = lt.getBoardFromFen(node.fen);
-      const mmw1 = this.maxMaterialWon(board, side) / 100;
-      const bril = (mmw1 - mmw3) * side - delta;
+      const bril = ((mmw3 - mmw1) * side - delta)*0.7; // TODO thesis: brilliant if there is a lot of material loss but still a good move
       return bril + bonus;
     };