+++ tools/preferences/tool.js
@@ -92,11 +92,13 @@
       const tools = lt.tools;
       const htmlEncode = lt.htmlEncode;
       const currentOptions = lt.currentOptions;
-      const applyOptions = lt.applyOptions.bind(lt);
+      const applyOptions = lt.saveOptions.bind(lt);
       const lichess = lt.lichess;
       const isOptionSet = lt.isOptionSet;
       const isLoggedIn = !!lt.getUserId();
 
+      const container = $('div.page-menu__content');
+      container.empty();
       lt.global.document.title = trans.noarg('lichessToolsPreferences');
 
       const subnav = $('main nav.subnav').hide();
@@ -301,10 +303,9 @@
 
       const prevFilter = $('input.prefFilter').val();
 
-      const container = $('div.page-menu__content');
       let saved = $('p.saved', container);
       saved = saved.length ? saved.clone() : $('<p class="saved text none">').attr('data-icon',lt.icon.Checkmark).text(trans.noarg('preferencesSaved'));
-      container.empty()
+      container
         .append(html)
         .addClass('lichessTools-preferences');